requirements.txt
psutil>=5.9.8
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
grpcio>=1.40.0
grpcio-tools>=1.40.0
pydantic>=1.8.0
msgspec>=0.18.0
requests>=2.25.0
httpx>=0.24.0
pytest-asyncio>=0.21
pyOpenSSL>=21.0.0
watchdog>=3.0.0
kubernetes>=26.1.0
opentelemetry-exporter-otlp-proto-http>=1.33.1
opentelemetry-exporter-otlp-proto-grpc>=1.33.1
pyzmq>=25.0.0
# Async event loop & concurrency
uvloop>=0.19.0
anyio>=4.0.0
# Process title (scheduler/worker process naming, optional at runtime)
setproctitle>=1.3.0
confluent_kafka>=2.13.2
transformers>=4.57.6
